% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_counts.R
\name{format_counts}
\alias{format_counts}
\title{Coerce mutation spectra matrix output from Helmsman into formats 
compatible with existing mutation signature analysis packages}
\usage{
format_counts(M, package)
}
\arguments{
\item{M}{Nx96 mutation count matrix output from Helmsman. Each row is a sample
ID, and each column is a trinucleotide mutation subtype. Each entry is the 
count of how many times a mutation with that trinucleotide context is seen 
in the sample.}

\item{package}{Name of mutation signature analysis package}
}
\value{
A Nx96 data frame with column and row names formatted according 
  to the specifications used
  Note that the SomaticSignatures packages uses a transposed mutation spectra 
  matrix, with subtypes as row names and sample IDs as column names
}
\description{
Coerce mutation spectra matrix output from Helmsman into formats 
compatible with existing mutation signature analysis packages
}
\examples{
\dontrun{
mu_counts <- read.table("/path/to/subtype_count_matrix.txt", header=T, stringsAsFactors=F)
msm_deconstructSigs <- format_counts(mu_counts, "deconstructSigs")
msm_SomaticSignatures <- format_counts(mu_counts, "SomaticSignatures")
msm_signeR <- format_counts(mu_counts, "signeR")
msm_maftools <- format_counts(mu_counts, "maftools")
msm_MutationalPatterns <- format_counts(mu_counts, "MutationalPatterns")
}
}
