% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remix_subtypes.R
\name{hcols_to_ds_cols}
\alias{hcols_to_ds_cols}
\title{Convert Helmsan subtype format to deconstructSigs subtype format}
\usage{
hcols_to_ds_cols(subtypes)
}
\arguments{
\item{subtypes}{A 96-element vector of mutation subtypes formatted as "C_A.TCG"}
}
\value{
A 96-element vector of mutation subtypes formatted as "T[C>A]G"
}
\description{
Convert Helmsan subtype format to deconstructSigs subtype format
}
\examples{
\dontrun{
mu_counts <- read.table("/path/to/subtype_count_matrix.txt", header=T, stringsAsFactors=F)
subtypes_in <- colnames(mu_counts)[-1]
signeR_rows <- hcols_to_ds_cols(subtypes_in)
}
}
