% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subtype_count_matrix}
\alias{subtype_count_matrix}
\title{Mutation spectra matrix for 2504 samples from the 1000 Genomes Project}
\format{A data frame with 2504 rows and 97 variables}
\source{
\url{http://www.diamondse.info/}
}
\usage{
subtype_count_matrix
}
\description{
A dataset containing the frequencies of 96 3-mer subtypes
}
\keyword{datasets}
