% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subtypes_bracket}
\alias{subtypes_bracket}
\title{Bracket-formatted subtypes}
\format{A 96-element vector of mutation subtypes formatted as "T[C>A]G"}
\usage{
subtypes_bracket
}
\description{
A dataset containing the frequencies of 96 3-mer subtypes
}
\keyword{datasets}
