% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subtypes_ss}
\alias{subtypes_ss}
\title{SomaticSignatures-formatted subtypes}
\format{A 96-element vector of mutation subtypes formatted as "CA T.G"}
\usage{
subtypes_ss
}
\description{
A dataset containing the frequencies of 96 3-mer subtypes
}
\keyword{datasets}
