% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subtypes_yapsa}
\alias{subtypes_yapsa}
\title{YAPSA-formatted subtypes}
\format{A 96-element vector of mutation subtypes formatted as "C>A TCG"}
\usage{
subtypes_yapsa
}
\description{
A dataset containing the frequencies of 96 3-mer subtypes
}
\keyword{datasets}
